+++ src/components/functional/wallet-connect-components/wallet-connect-modal.types.ts
@@ -1,9 +1,0 @@
-export enum WalletConnectEventsEnum {
-  'CLOSE' = 'CLOSE', //  close modal event
-  'DATA_UPDATE' = 'DATA_UPDATE', // update data event
-}
-
-export interface IWalletConnectModalData {
-  wcURI: string;
-  shouldClose?: boolean;
-}